+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Add `#rails_routes` to Roda and Rodauth instance for accessing Rails route helpers (@janko)
+
 * Add `#rails_request` to Roda and Rodauth instance for retrieving an `ActionDispatch::Request` instance (@janko)
 
 ## 0.10.0 (2021-03-23)